+++ Calculator/src/main/java/Application/Calculator/Calculator.java
@@ -11,7 +11,7 @@ public class Calculator {
      * @param b druga liczba
      * @return suma dwóch liczb
      */
-    public int add(int a, int b) {
+    public double add(double a, double b) {
         return a + b;
     }
 
@@ -22,7 +22,7 @@ public class Calculator {
      * @param b druga liczba
      * @return różnica dwóch liczb
      */
-    public int subtract(int a, int b) {
+    public double subtract(double a, double b) {
         return a - b;
     }
 
@@ -33,7 +33,7 @@ public class Calculator {
      * @param b druga liczba
      * @return iloczyn dwóch liczb
      */
-    public int multiply(int a, int b) {
+    public double multiply(double a, double b) {
         return a * b;
     }
 
@@ -45,7 +45,7 @@ public class Calculator {
      * @return wynik dzielenia dwóch liczb
      * @throws IllegalArgumentException jeśli dzielnik wynosi zero
      */
-    public int divide(int a, int b) throws IllegalArgumentException {
+    public double divide(double a, double b) throws IllegalArgumentException {
         if (b == 0) {
             throw new IllegalArgumentException("Nie można dzielić przez zero!");
         }
